pyproject.toml
[project]
name = "sentcite"
version = "0.1.0"
description = "Sentence-level citation RAG prototype for tax audit workflows."
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
  "pydantic>=2.6",
  "python-dotenv>=1.0",
  "azure-identity>=1.17",
  "azure-storage-blob>=12.20",
  "azure-ai-documentintelligence>=1.0.0",
  "azure-search-documents>=11.6.0b4",
  "openai>=1.40",
  "spacy>=3.7",
  "numpy>=1.26",
  "scikit-learn>=1.4",
  "tiktoken>=0.7",
  "pyyaml>=6.0",
  "tqdm>=4.66",
  "rich>=13.7",
]

[project.optional-dependencies]
dev = [
  "pytest>=8.0",
  "pytest-cov>=5.0",
  "ruff>=0.5",
  "jupyter>=1.0",
  "ipykernel>=6.29",
]

[build-system]
requires = ["setuptools>=68"]
build-backend = "setuptools.build_meta"

[tool.setuptools.packages.find]
where = ["src"]

[tool.pytest.ini_options]
testpaths = ["tests"]
addopts = "-ra -q"

[tool.ruff]
line-length = 100
target-version = "py311"
